Add sidebar spec for new project issue boards (EE)
- Remove the stub for `graphql_board_lists` from sidebar_spec.rb (EE).

This adds spec coverage for project issue boards when it is using
graphql-based boards code. A few selectors are added for testing
in sidebar components.

- Create a temporary spec for the classic boards

We will simply clone `sidebar_spec.rb` as `sidebar_deprecated_spec.rb`
and continue to provide the spec coverage for project issue boards (EE)
when it is using the classic boards code.