Remove feature flag for LFS integrity check via quarantined objects
The change to use the quarantine environment for LFS integrity checks
has been enabled in production now for some time with the expected
benefits: if used, the new check is up to several magnitudes faster than
the old LFS pointer check.

Let's drop the feature flag and make this new feature generally
available.