Add ability to include self in cascading setting lock check
Currently, calling namespace_settings "#{attribute}_locked?" will
return `false` if not locked by an ancestor, but if it is locked
by the current level. This adds the ability to specify that the
current level should be included, which is useful when looking
at locks from the project level, rather than the group level.

Changelog: changed