Fix N+1 in Releases API entity generation
Author and links count queries happened per release, added them
to includes to preload this and remove the N+1.

The links are exposed as links.sorted in the API entity, in order
to preload it, a scope that includes the sort had to be created,
and that included in the preloaded scope to prevent sorting links
for every release as an N+1.