API caching for Releases
Introduces output caching for the Releases API.

Rendering the release entities and dumping the JSON seems to take most
of the time on this endpoint, which should be removed with caching.

Cache invalidation happens on project & user & permissions, important
as the user can have hierarchical access through shared groups, and
some of the data is hidden depending on this.

Changelog: performance