Reorder source_project_id foreign key constraint
Project deletion was failing due to a SQL timeout on
setting `source_project_id` to `nil` in projects table.

One of the reason why this was happening is:

For a non-forked project, the source_project_id and target_project_id
will be same as the project in which the MR is created.
`source_project_id` has the condition ON DELETE SET NULL and
`target_project_id` has the condition ON DELETE CASCADE.

When the project gets deleted, the source_project_id is
first set to `nil` and then the cascade deletion happens
and all records are deleted.

This occurs because the constraint for `source_project_id` is defined
earlier than the `target_project_id`.

This MR fixes this issue by re-creating the foreign key
so that the `target_project_id` constraint is placed before
`source_project_id`.