Speed up Elastic AddNewDataToMergeRequestsDocuments migration
This migration was just introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/59836 .

After watching this migration running in production we can see that we
can make it run faster. The numbers here also match what we used in
`AddPermissionsDataToNotesDocuments` so should be safe.

I've also changed it to use the `ProcessInitialBookkeepingService` which
is basically just moving the work into a different queue. We have 2
queues for indexing work in Elasticsearch and the "initial" queue is for
newly created projects/imported projects. It gets much less traffic than
the so called "incremental" queue and therefore it's safer to put
backfill migrations into this queue. We actually should have used this
queue to begin with but switching it now is perfectly fine.

Changelog: changed