Move scanner warnings in Vulnerability Report
This moves the scanner warning alert messages to the top of the Project
Vulnerability Report, rather than in the first row of the vulnerability
table.

This also moves the _logical_ location of the alert from the
`VulnerabilityList` component to the `ProjectVulnerabilities` component,
since the data that determines whether to show the alert is fetched by
the `ProjectVulnerabilities` component, and not the `VulnerabilityList`.

The alert can only be shown in the context of a project's Vulnerability
Report anyway, but the `VulnerabilityList` is also used in the content
of a group and the whole GitLab instance. Therefore, it made sense to
hoist this logic up to the project-specific component.

Finally, the visual positioning of the alert is achieved via portal-vue.
This is added as a new top-level dependency in GitLab, but in fact it
was already a transient dependency via GitLab UI

Addresses https://gitlab.com/gitlab-org/gitlab/-/issues/322889.

Changelog: changed