Fix alert management fingerprints
ActionController::Parameters class is now prefixed with `#`
and we have to remove the symbol in order to have the same
fingerpring for the existing alerts with the same payload

This commit can be extracted out of the main MR and discussed
with the dedicated team how we can better fix the issue.

One of the solutions would be relying on the .to_json output
instead of .to_s, because its format is unlikely to change.


(cherry picked from commit aa44cd514b58dbbe392408e2cd3a1c438a484562)