Update lock timeout in ElasticCommitIndexerWorker
This lock prevents indexing the same project twice in parallel which is
a big problem. The lock will clear itself when the job finishes but you
need a timeout since the lock lives in Redis and the process may crash
before it can clear the lock.

We chose 1 hour originally since we assumed this was enough time to
finish indexing any project but we've found from several customer
tickets https://gitlab.com/gitlab-org/gitlab/-/issues/323856 that some
very large projects are taking longer than an hour to finish indexing.
Increasing this to 1 day probably will dramatically reduce the risk of 2
projects indexing at the same time while still providing a reasonable
timeout in the rare event the process crashes and can't clean up the
lock.