Keep new prometheus cluster integration in sync with app
This is the next step of the Prometheus cluster application deprecation.
It is the "double-write" phase of a table data migration, with the
cluster application as the source of truth, and all changes mirrored by
the cluster integration. Once this is live on GitLab.com, we can perform
a data migration to backfill the cluster integrations and start using
them as the source of truth for the "enabled" status for the in-cluster
prometheus functionality.

See https://gitlab.com/gitlab-org/gitlab/-/issues/327655