Initial namespacing services in `Integrations`
This change moves 3 models in `project_services` to the `Integrations`
namesapce as part of
https://gitlab.com/gitlab-org/gitlab/-/issues/201855.

It represents an initial step in the process, aiming to allow the STI
classes to continue to function as before with the use of a new customer
attribute type (adapter) `Gitlab::Integrations::StiType`. This adapter
can be used to selectively move models into the new namespace while
maintaing the legacy STI `type` values (e.g `AsanaService`) needed in
order for the STI modelling to continue to work.

When all models have been namespaced, we could consider then performing
data migrations of the `type` values, and eventually remove the
`StiType` altogether.