Add more attributes to the blob GraphQL API
`ide_edit_path` and `can_modify_blob` both have implementations in
`BlobHelper`, which are difficult to extract out. For now, just include
the required helpers to get access to the code.

In the future, we should be able to delete the existing FE for viewing
the repository, and at that point we should also be able to move this
code from BlobHelper and into BlobPresenter more easily.

`external_storage_url` has its implementation in `ApplicationHelper`.
Long term, moving this seems less likely. It probably fits better in a
URL generation location like `GitlabRoutingHelpers`, but let's not try
to tackle that as part of this MR.

Changelog: added