checks: Refactor EE checks to receive changes as argument
We're refactoring access checks such that they not only have
per-reference checks, but such that they also have batched checks which
go over all changes at once. This new architecture doesn't match with
the current design of our Checks, which all host a single change as
member variables and thus are hard to reuse.

Refactor the code to explicitly accept the change as parameter.