Move snippets check to after restoring from backup
Once backups and restores are handled by gitaly-backup we will not be
able to modify the gitlab database during a backup or restore. To work
around this we need to move all such checks to after the process has
completed. Right now this includes: validating snippets, and
rescheduling repository pool creation.