Refactor Backup::Repositories specs
Now that we have extracted the gitaly RPC backup strategy and tested it
separately, we can make the tests more generic. This makes sure we are
querying the database and calling the strategy correctly, but not trying
to actually verify how the backup happens. The goal here is to make the
strategy itself the single source of truth and not make these tests too
fragile.