Backfill clusters_integration_elastic_stack.enabled
We are in the process of deprecating GitLab-managed apps feature in
GitLab, in favour of user-managed integrations. This migration backfills
the integration status based on the corresponding application status.

With this migration completed, we can start using the integration record
as the source of truth for whether to enable the integration.

Second part of https://gitlab.com/gitlab-org/gitlab/-/issues/326560

Changelog: changed