Mount vulnerability list on pipeline dashboard
This is part of an ongoing development process. It mounts the vulnerability
list and provides the basic functionality like displaying vulnerabilities
and fetching more. Currently the sorting and filtering does not work yet.
Subsequent MRs will take care of that functionality. This feature is
being developed behind a feature flag.