Remove User Lists from Feature Flag Table
The table of user lists will be moving to its own page, so the first
step is to clean up some of the feature flag table. Here, we remove the
tab showing the user lists, as well as the component's ability to be in
a tab.

This is only the beginning.