Create User Lists Page and Route
Also move all of the user list table over to its new home!

This is mostly to keep the two concepts separated, and allow for easier
iteration on the feature flags table. I want to extract the single
shared component, but for now it is easy to keep in feature flags.

Changelog: added