Fix animation for pipeline editor drawer on load
When the page was loading, we had a transition all
which resulted in the drawer being moved diagonally
into position and it looked weird. Instead, we now
only transition on width property. Also, we default
the state of the drawer to close and we
programatically set it to open if there is no
value in the local storage. That way, we get
a nice opening animation that wasn't there before.