Deduplicate ManageViaMr specs
Previously the FOSS and EE specs for the ManageViaMr component were very
nearly identical. The only difference between them was a different
parameterisation via `describe.each`.

This removes the EE spec, and makes the FOSS spec rely on the
`ee_else_ce` import alias to switch out the appropriate parameters, such
that the correct tests are run in FOSS and EE modes in CI.