Modify sentry for big pipeline graph
Because we were using `captureError`
instead of `captureMessage` for the error
handling in sentry, the hard char limit
was 200 char instead of over 8k. This
commit changes that for an error in which
we need a lot of char. Also, refactors
the parsing util for errors since
it was not working properly.