Fix job argument serialization for deduplication
Sidekiq uses `Sidekiq.dump_json` and `Sidekiq.load_json` for
serializing and deserializing jobs and their arguments in Redis.
This supports entire objects being serialized into JSON.

Before this, we were counting on an array of primitives as the
arguments and serialized it naively to determine an idempotency
key. This would cause the deletion of the key in Redis to fail in
Sidekiq, because we'd be trying to delete a different key than the
client had written.

For example:

```
[29, 1, [], { execute_hooks: true }].join('-')

[29, 1, [], Sidekiq.load_json(Sidekiq.dump_json({ execute_hooks: true }))].join('-')
```

With this, we're serializing the keys on the client-side the same way
the server would, so we'd be talking about the same keys to delete.

Changelog: fixed