Fix double render in project's git URL redirect
When you visit the git URL of a project using a browser, we redirect to
the project's URL without the git suffix.

When a project is moved, visiting the old URL results in a 500 because
we're redirecting twice in a single before_action.

Changelog: fixed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/62053