Bugfix `@service` being nil in integrations views
In 5fd1c950e3e3ea5ad58cd2ad3e6c0f011874ebd2, `@service` was renamed
in `IntegrationsActions` (used by admin and group integrations
controllers) to `@integration`. We need to continue to support that
variable as it is referred to in shared views.

This fix applies the same approach as in the original commit to continue
to support this variable:
https://gitlab.com/gitlab-org/gitlab/-/blob/5fd1c950e3e3ea5ad58cd2ad3e6c0f011874ebd2/app/controllers/admin/services_controller.rb#L39

Removing the need for this will be addressed in:
https://gitlab.com/gitlab-org/gitlab/-/issues/329759

https://gitlab.com/gitlab-org/gitlab/-/issues/331262