Don't record duration for failing requests
If a request failed, we shouldn't care how fast it failed.

These recordings are used for the apdex aspect for error budget
recordings for stage groups. With the current implementation, if a
stage group has a single request that failed fast, they would have 50%
availability. This should be 0% to match how we measure availability
for the services themselves.

For those, we only measure apdex if the request succeeded:
https://gitlab.com/gitlab-org/gitlab/blob/d969d80acb4bb416d2b28423de6810d314a3cfc8/lib/gitlab/metrics/requests_rack_middleware.rb#L82

Changelog: fixed