Move cluster management projects out of alpha
With the removal of one-click applications from the UI in 14.0, and the
deprecation of GitLab-managed applications, users are expected to manage
their own Helm releases.

Users are free to do this however they want, but to facilitate
infrastructure-as-code practices, we are moving Cluster Management
Projects out of alpha.

See https://gitlab.com/groups/gitlab-org/-/epics/4280

Changelog: changed