Reduce number of queries for finding iterations
Reduce number of queries made when finding iterations, that is
due to the fact that iterations are accessible throughout the
group hierarchy and not available at project level. So we do not
need to check for project iterations and it is eniough to check
that user can read iterations at the hierarchy leaf group.