Remove misleading table definitions
Some users may read the table docs as
indicating that the width of table cells
can be controlled by putting more or less
hyphens in the second row.
This isn't true, but it's implied -
slightly - by the formatting of our example
tables.
Combined with phrases like "The cell sizes
**don't** have to match each other. They are
flexible", users may infer that cell sizes
are **controllable**.

This change eliminates this ambiguity.