Handle enter action of MR comment form
When you have draft review comments in an MR, the primary action
button of the main comment form becomes "Add to review".
However, when you do cmd+enter in the comment form, it adds the comment
right away. The secondary button is triggered. With this comment,
the primary button will work when you do cmd+enter.

Changelog: changed