Add speedscope profiling to Request Profiles
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/53288 added
Speedscope output to the performance bar, which is very useful for
profiling requests in the browser. However, this cannot be used to
profile API requests.

This commit adds a 'speedscope' profile type to the existing 'Request
Profiles' system, which allows both browser and API calls to be profiled
by setting the 'X-Profile-Token' header on requests.