Set a flag to allow unsupported election strategy in Praefect
Praefect has deprecated other election strategies than 'per_repository'
in 13.12. The legacy strategies are scheduled to be removed in 14.0.
GitLab's tests are currently using one of them. Updating the tests to
support the new strategy is slightly involved as the new strategy
requires a database. To allow for Gitaly team to prevent the use of the
legacy election strategies from 14.0 onwards but to keep the GitLab tests
still passing before the setup can be reworked, we added a configuration
flag to explicitly allow these unsupported election strategies to be
configured still. This commit sets the flag for GitLab's test so they
keep working for now until we can set up a database connection in them.