Document new dataloss behavior following Praefect's failover changes
Praefect's dataloss now includes in the output repositories which are
unavailable on the cluster. With variable replication factor, it's no
longer sufficient to only list the repositories which have outdated,
assigned replicas. A repository with replication factor of one may be
fully up to date but may be unavailable if the only replica is unhealthy.
This amounts to dataloss if the Gitaly node is not brought back online.

Praefects lazy failover behavior has also made the read-only mode non-existent,
as Praefect now immediately fails over to an up to date replica if one
is available. Praefect also no longer elects outdated replicas as primaries.
Given these changes, this commit updates the dataloss documentation to
reflect these behaviors.