Use cluster integrations by default
We are removing one-click applications from GitLab but leaving the
integrations for Prometheus and Elastic Stack.

We have code in place that keeps the integration in sync with the
corresponding cluster application, and have backfilled the
ingtegrations, so we can now use them as a source of truth.

See

- Backfill Prometheus: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/61502
- Backfill Elastic Stack: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/61521
- Deprecation Epic: https://gitlab.com/groups/gitlab-org/-/epics/4280

Changelog: changed