Show alert if required for the user
We show credit card validation alert for
users when their pipeline failed and the user is not verified.
That's not enough because we now have a special setting in the admin
panel for toggling requirement of the credit card validation.

Changelog: changed
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/62735