Modify index on members table
Modify the index_non_requested_project_members_on_source_id_and_type
index. Since we can't actually modify the index, we need to remove
and add it back with the modified options.

Modify the index to use Postgres's include feature, to include
additional columns that allow the index to be used in the
Group#billed_project_members query.

See
https://gitlab.com/gitlab-org/gitlab/-/issues/329936#note_581777031